Fix possible loss of data compilation error: lastmod_failedToFindFile
omrfile_lastmod returns an int64_t. In the lastmod_failedToFindFile
testcase, the return value from omrfile_lastmod is stored in an int.
This causes the following error on Windows:

omrfileTest.cpp(4801) : error C2220: warning treated as error - no
'object' file generated
omrfileTest.cpp(4801) : warning C4244: 'initializing' : conversion from
'int64_t' to 'int', possible loss of data

In the above testcase, the return value is stored in an int64_t to
resolve the compilation error.

+++ fvtest/porttest/omrfileTest.cpp
@@ -4798,7 +4798,7 @@ TEST_F(PortFileTest2, lastmod_failedToFindFile)
 	const char *fileName = "lastmod_failedToFindFile.tst";
 	omrfile_unlink(fileName);
 
-	int rc = omrfile_lastmod(fileName);
+	int64_t rc = omrfile_lastmod(fileName);
 	if (rc != -1) {
 		outputErrorMessage(PORTTEST_ERROR_ARGS, "omrfile_lastmod() did not return -1 on an invalid file.");
 	}
